server/requirements.txt
fastapi~=0.121.2
uvicorn[standard]
python-multipart

pydantic~=2.12.4
pydantic-settings~=2.12.0

google-genai
python-dateutil~=2.9.0.post0
dateparser

langgraph
langchain
langchain-core
langchain-community

jinja2
rich

pytest
pytest-asyncio
httpx

